Refactor GeoIP classes
