+++ mirror-map/pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.warvan1</groupId>
-    <artifactId>mirrormap</artifactId>
+    <groupId>edu.clarkson.cosi</groupId>
+    <artifactId>mirror-map</artifactId>
     <version>0.0.1</version>
 
     <properties>
