+++ mirror-map/src/main/java/mirrormap/MirrorMapApplication.java
@@ -1,12 +1,7 @@
 package mirrormap;
 
 import java.io.IOException;
-import java.lang.Thread;
 
-import com.maxmind.geoip2.exception.GeoIp2Exception;
-
-import mirrormap.maxmind.DatabaseHandler;
-import mirrormap.maxmind.DatabaseUpdater;
 import mirrormap.server.ServerController;
 
 public class MirrorMapApplication {
